Issue #412 [Blast Search] Improve BLAST task listing method

+++ server/catgenome/src/main/java/com/epam/catgenome/manager/blast/dto/BlastRequest.java
@@ -14,7 +14,7 @@ public class BlastRequest {
     private List<Long> taxIds;
     private List<Long> excludedTaxIds;
     private String query;
-    private Long maxTargetSequence;
-    private Long expectedThreshold;
+    private long maxTargetSequence;
+    private double expectedThreshold;
     private String options;
 }
